// todo-service/src/utils/http-status-code.enum.ts
export enum HttpStatusCode {
  OK = 200,
  BAD_REQUEST = 400,
  UNAUTHORIZED = 401,
  PAYMENT_REQUIRED = 402,
  FORBIDDEN = 403,
  NOT_FOUND = 404,
  CONFLICT = 409,
  PAYLOAD_TOO_LARGE = 413,
  INVALID = 416,
  PAYMENT_ERROR = 420,
  LIMIT_EXCEEDED = 429,
  EXPIRED = 444,
  MISSING_TRANSACTION_TYPE = 455,
  FILE_NOT_FOUND = 450,
  NOT_RECOGNIZE = 451,
  MISSING_ADDRESS = 472,
  CONSISTENT = 476,
  BLOCKED = 477,
  UNHANDLED_REJECTION = 478,
  INTERNAL_SERVER = 500,
  NOT_IMPLEMENTED = 501,
  SERVICE_UNAVAILABLE = 502,
}